+++ internal/services/payment.service.go
@@ -9,7 +9,6 @@ import (
 	"github.com/2110366-2566-2/tortoise-app-backend/internal/database"
 	"github.com/2110366-2566-2/tortoise-app-backend/internal/models"
 	"github.com/gin-gonic/gin"
-	"github.com/stripe/stripe-go/v72"
 	"github.com/stripe/stripe-go/v72/paymentintent"
 	"go.mongodb.org/mongo-driver/bson"
 )
@@ -26,21 +25,6 @@ func NewPaymentHandler(handler *database.Handler, env configs.EnvVars) *PaymentH
 	}
 }
 
-func logStripeError(err error) {
-	if stripeErr, ok := err.(*stripe.Error); ok {
-		switch stripeErr.Type {
-		case stripe.ErrorTypeCard:
-			log.Println("A payment error occurred:", stripeErr.Msg)
-		case stripe.ErrorTypeInvalidRequest:
-			log.Println("An invalid request occurred.")
-		default:
-			log.Println("Another Stripe error occurred.")
-		}
-	} else {
-		log.Println("An error occurred that was unrelated to Stripe.")
-	}
-}
-
 // CreatePayment godoc
 // @Method POST
 // @Summary Create payment
@@ -52,43 +36,29 @@ func (h *PaymentHandler) CreatePayment(c *gin.Context) {
 	if err := c.ShouldBindJSON(&transaction); err != nil {
 		log.Println("failed to bind JSON: ", err)
 		c.JSON(400, gin.H{"success": false, "error": "failed to bind JSON"})
-		return // Return here
+		return
 	}
 
 	isSold, err := h.handler.CheckPetStatus(c, transaction.PetID)
 	if err != nil {
 		log.Println("failed to check pet status: ", err)
 		c.JSON(400, gin.H{"success": false, "error": "failed to check pet status"})
-		return // Return here
+		return
 	}
 
 	if isSold {
 		log.Println("pet is not available")
 		c.JSON(400, gin.H{"success": false, "error": "pet is not available"})
-		return // Return here
+		return
 	}
 
-	// Set Stripe API key
-	stripe.Key = h.env.STRIPE_KEY
-
-	params := &stripe.PaymentIntentParams{
-		Amount:             stripe.Int64(transaction.Price * 100),
-		Currency:           stripe.String(string(stripe.CurrencyTHB)),
-		PaymentMethodTypes: stripe.StringSlice([]string{"card", "promptpay"}),
-	}
-
-	// Add metadata
-	params.AddMetadata("pet_id", transaction.PetID.Hex())
-	params.AddMetadata("seller_id", transaction.SellerID.Hex())
-	params.AddMetadata("buyer_id", transaction.BuyerID.Hex())
-
-	pi, err := paymentintent.New(params)
+	pi, err := h.CreatePaymentIntent(transaction)
 	if err != nil {
-		logStripeError(err)
 		// log.Println("failed to create transaction: ", err)
 		c.JSON(400, gin.H{"success": false, "error": "failed to create transaction"})
-		return // Return here
+		return
 	}
+
 	// Update pet status
 	_, err = h.handler.UpdatePetStatus(c, transaction.PetID, true)
 
@@ -99,20 +69,16 @@ func (h *PaymentHandler) CreatePayment(c *gin.Context) {
 		_, err := paymentintent.Cancel(pi.ID, nil)
 		if err != nil {
 			c.JSON(400, gin.H{"success": false, "error": "failed to cancel transaction"})
-			return // Return here
+			return
 		}
 		c.JSON(400, gin.H{"success": false, "error": "failed to create transaction"})
-		return // Return here
+		return
 	}
 
-	// Add transaction ID to transaction model
+	// Add transaction ID, payment status, and timestamp
 	transaction.PaymentID = pi.ID
-
-	// Add timestamp
-	transaction.Timestamp = time.Now()
-
-	// Add transaction status
 	transaction.Status = "pending"
+	transaction.Timestamp = time.Now()
 
 	// Add transaction to database
 	_, err = h.handler.CreateTransaction(c, &transaction)
@@ -122,25 +88,26 @@ func (h *PaymentHandler) CreatePayment(c *gin.Context) {
 		// rollback
 		_, err1 := h.handler.UpdatePetStatus(c, transaction.PetID, false)
 		if err1 != nil {
-			c.JSON(400, gin.H{"success": false, "error": "failed to rollback"})
-			return // Return here
+			fmt.Println("failed to rollback: ", err1)
+			c.JSON(400, gin.H{"success": false, "error": "failed to create transaction"})
+			return
 		}
 		// cancel transaction intent
 		_, err2 := paymentintent.Cancel(pi.ID, nil)
 		if err2 != nil {
-			c.JSON(400, gin.H{"success": false, "error": "failed to cancel transaction"})
-			return // Return here
+			fmt.Println("failed to cancel transaction: ", err2)
+			c.JSON(400, gin.H{"success": false, "error": "failed to create transaction"})
+			return
 		}
+		fmt.Println("failed to create transaction: ", err)
 		c.JSON(400, gin.H{"success": false, "error": "failed to create transaction"})
-		return // Return here
+		return
 	}
 
-	res := bson.M{
+	c.JSON(201, gin.H{"success": true, "data": bson.M{
 		"transaction_id": transaction.ID.Hex(),
 		"payment_id":     pi.ID,
-	}
-
-	c.JSON(200, gin.H{"success": true, "data": res})
+	}})
 }
 
 // ConfirmPayment godoc
@@ -164,6 +131,7 @@ func (h *PaymentHandler) ConfirmPayment(c *gin.Context) {
 	res, err := h.handler.UpdateTransaction(c, payment.TransactionID, update)
 
 	if err != nil {
+		fmt.Println("failed to update transaction")
 		c.JSON(400, gin.H{"success": false, "error": "failed to update transaction"})
 		return // Return here
 	}
@@ -176,28 +144,10 @@ func (h *PaymentHandler) ConfirmPayment(c *gin.Context) {
 		return
 	}
 
-	// Set Stripe API key
-	stripe.Key = h.env.STRIPE_KEY
-
-	var PaymentMethod string
-
-	if transaction.PaymentMethod == "promptpay" {
-		// PaymentMethod = "promptpay"
-		PaymentMethod = "pm_card_th_credit"
-	} else {
-		PaymentMethod = "pm_card_th_credit"
-	}
-
-	_, err = paymentintent.Confirm(
-		payment.ID,
-		&stripe.PaymentIntentConfirmParams{
-			PaymentMethod: stripe.String(PaymentMethod),
-		},
-	)
+	err = h.ConfirmPaymentIntent(&transaction, &payment)
 
 	// Check if there is an error
 	if err != nil {
-		logStripeError(err)
 		// rollback
 		_, err = h.handler.UpdateTransaction(c, payment.TransactionID, bson.D{{Key: "$set", Value: bson.D{{Key: "status", Value: "failed"}}}})
 		if err != nil {
