+++ internal/transport/rest/apiV1/pet.transport.go
@@ -1,0 +1,33 @@
+package apiV1
+
+import (
+	"github.com/2110366-2566-2/tortoise-app-backend/internal/database"
+	"github.com/2110366-2566-2/tortoise-app-backend/internal/services"
+	"github.com/gin-gonic/gin"
+)
+
+func PetController(r *gin.RouterGroup, h *database.Handler) {
+	// Create a new pet handler
+	petHandler := services.NewPetHandler(h)
+
+	allUser := r.Group("/")
+	allUser.Use(roleMiddleware("seller", "admin", "buyer"))
+
+	sellerAdmin := r.Group("/")
+	sellerAdmin.Use(roleMiddleware("seller", "admin"))
+
+	// Set up routes
+	allUser.GET("/", petHandler.GetAllPets)
+	allUser.GET("/filter", petHandler.GetFilteredPets)
+	allUser.GET("/:petID", petHandler.GetPetByPetID)
+	allUser.GET("/seller/:userID", petHandler.GetPetBySeller)
+
+	sellerAdmin.POST("/seller/:userID", petHandler.CreatePet)
+	sellerAdmin.PUT("/:petID", petHandler.UpdatePet)
+	sellerAdmin.DELETE("/:petID", petHandler.DeletePet)
+
+	// master data
+	allUser.GET("/master", petHandler.GetMasterData)
+	allUser.GET("/master/:category", petHandler.GetMasterDataByCategory)
+	allUser.GET("/master/categories", petHandler.GetCategories)
+}
